+++ tests/testthat/test-server.R
@@ -145,7 +145,6 @@ test_that("validate baseline", {
   response <- response_from_json(r)
   expect_equal(response$status, "success")
   expect_equal(response$errors, list())
-  expect_equal(response$data$complete, TRUE)
   expect_equal(response$data$consistent, TRUE)
 })
 
@@ -438,6 +437,8 @@ test_that("spectrum file download streams bytes", {
     r <- httr::GET(paste0(server$url, "/download/spectrum/", response$data$id))
     expect_equal(httr::status_code(r), 200)
     expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+    expect_match(httr::headers(r)$`content-disposition`,
+                 'attachment; filename="naomi_spectrum_digest_\\w+.zip"')
     ## Size of bytes is close to expected
     size <- as.numeric(httr::headers(r)$`content-length`)
     expect_true(size - size/10 <
@@ -445,6 +446,13 @@ test_that("spectrum file download streams bytes", {
     expect_true(size + size/10 >
       file.size(system_file("output", "malawi_spectrum_download.zip")))
   })
+
+  ## Headers can be retrieved
+  r <- httr::HEAD(paste0(server$url, "/download/spectrum/", response$data$id))
+  expect_equal(httr::status_code(r), 200)
+  expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+  expect_match(httr::headers(r)$`content-disposition`,
+    'attachment; filename="naomi_spectrum_digest_\\w+.zip"')
 })
 
 test_that("summary file download streams bytes", {
@@ -468,6 +476,8 @@ test_that("summary file download streams bytes", {
     r <- httr::GET(paste0(server$url, "/download/summary/", response$data$id))
     expect_equal(httr::status_code(r), 200)
     expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+    expect_match(httr::headers(r)$`content-disposition`,
+                 'attachment; filename="naomi_summary_\\w+.zip"')
     ## Size of bytes is close to expected
     size <- as.numeric(httr::headers(r)$`content-length`)
     expect_true(size - size/10 <
@@ -475,6 +485,13 @@ test_that("summary file download streams bytes", {
     expect_true(size + size/10 >
       file.size(system_file("output", "malawi_summary_download.zip")))
   })
+
+  ## Headers can be retrieved
+  r <- httr::HEAD(paste0(server$url, "/download/summary/", response$data$id))
+  expect_equal(httr::status_code(r), 200)
+  expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+  expect_match(httr::headers(r)$`content-disposition`,
+               'attachment; filename="naomi_summary_\\w+.zip"')
 })
 
 test_that("can quit", {
@@ -491,4 +508,35 @@ test_that("can quit", {
   expect_is(r, "error")
 
   expect_false(server$process$is_alive())
+})
+
+test_that("404 pages have sensible schema", {
+  server <- hintr_server()
+  r <- httr::GET(paste0(server$url, "/meaning-of-life"))
+  expect_equal(r$status_code, 404)
+  expect_equal(r$headers[["content-type"]], "application/json")
+
+  dat <- httr::content(r, "parsed", encoding = "UTF-8")
+  expect_equal(dat$status, "failure")
+  expect_equal(dat$errors[[1]]$error,
+               "NOT_FOUND")
+  expect_equal(dat$errors[[1]]$detail,
+               "GET /meaning-of-life is not a valid hintr path")
+})
+
+test_that("Error handler is triggered", {
+  ## This test is ugly because it probably should be fixed.
+  server <- hintr_server()
+  r <- httr::GET(paste0(server$url, "/download/summary/asdfasdfa"))
+  expect_equal(r$status_code, 500)
+  expect_equal(r$headers[["content-type"]], "application/json")
+
+  dat <- httr::content(r, "parsed", encoding = "UTF-8")
+  expect_equal(dat$status, "failure")
+  expect_equal(dat$errors[[1]]$error,
+               "SERVER_ERROR")
+  detail <- paste("Unexpected server error in '<call missing>' :",
+                  "'$ operator is invalid for atomic vectors' while doing",
+                  "'GET /download/summary/asdfasdfa'")
+  expect_equal(dat$errors[[1]]$detail, detail)
 })